// frontend/src/types/OKRs.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { OKRObjective } from './OKRObjective';
/**
 * Collection of objectives and key results
 */
export type OKRs = {
    /**
     * List of objectives with their key results
     */
    objectives: Array<OKRObjective>;
};